requirements.txt
crawl4ai
llama-index
qdrant-client
fastapi
uvicorn
python-dotenv
openai
pydantic
requests
pytest
